Se arregla pequeño fallo en el mapeo de las consultas de movimientos varios de las tarjetas de la ficha de certificaciones

+++ siga-core/src/main/java/org/itcgae/siga/DTOs/scs/MovimientosVariosApliCerItem.java
@@ -2,6 +2,7 @@ package org.itcgae.siga.DTOs.scs;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -17,7 +18,7 @@ public class MovimientosVariosApliCerItem {
     private String nombre;
     private String descripcion;
     private Date fechaAlta;
-    private Short importeAplicado;
+    private BigDecimal importeAplicado;
     private String asunto;
 }
